refactor(usecase-pattern): apply usecase pattern to food and measure modules
