+++ .serena/memories/usecase-pattern-migration-complete.md
@@ -1,0 +1,100 @@
+# Usecase Pattern Migration - Complete Implementation
+
+## Summary
+
+Successfully applied the new usecase pattern from day-diets module to all recently refactored modules in the codebase, following the standardized architecture established in recent commits.
+
+## Analysis Results
+
+### Modules Already Following Usecase Pattern
+
+1. **day-diet** ✅ (Reference Implementation)
+   - `usecases/dayChange.ts` - Day change operations  
+   - `usecases/dayCrud.ts` - CRUD operations
+   - `usecases/dayState.ts` - State management with re-exports
+
+2. **macro-profile** ✅ (Already Compliant)
+   - `usecases/macroProfileCrud.ts` - CRUD operations
+   - `usecases/macroProfileState.ts` - State management with re-exports
+
+3. **recent-food** ✅ (Already Compliant)  
+   - `usecases/recentFoodCrud.ts` - CRUD operations only
+
+4. **recipe** ✅ (Already Compliant)
+   - `usecases/recipeCrud.ts` - CRUD operations only
+
+### Modules Requiring Refactoring
+
+#### 1. **measure** Module - REFACTORED ✅
+
+**Before:**
+- Single file `measureCrud.ts` with mixed concerns (state + CRUD + realtime)
+
+**After:**
+- `usecases/measureCrud.ts` - Pure CRUD operations
+- `usecases/measureState.ts` - State management with re-exports and realtime initialization
+
+**Changes:**
+- Separated state management from CRUD operations
+- Moved `createResource` and `refetchBodyMeasures` to state file
+- Removed manual refetch calls from CRUD functions 
+- Updated all imports across codebase to use state file for reactive data
+
+#### 2. **food** Module - REFACTORED ✅
+
+**Before:**
+- Single file `food.ts` with all operations mixed together
+
+**After:**
+- `usecases/foodCrud.ts` - All CRUD operations
+- `food.ts` - Re-export file following the pattern
+
+**Changes:**
+- Moved all functions to dedicated CRUD file
+- Maintained all existing functionality and error handling
+- Updated main application file to re-export from usecases
+
+## Pattern Consistency
+
+All modules now follow the standardized usecase pattern:
+
+```
+application/
+├── usecases/
+│   ├── [module]Crud.ts    - Pure CRUD operations
+│   └── [module]State.ts   - State management + re-exports (if needed)
+└── [module].ts            - Re-export file (legacy compatibility)
+```
+
+## Quality Assurance
+
+- ✅ All TypeScript checks pass
+- ✅ All ESLint checks pass  
+- ✅ All tests pass (310/310)
+- ✅ No breaking changes to existing APIs
+- ✅ Maintained backward compatibility through re-exports
+
+## Architectural Benefits
+
+1. **Clear Separation of Concerns**: CRUD operations separated from state management
+2. **Consistent Patterns**: All modules follow the same organizational structure
+3. **Maintainability**: Easier to locate and modify specific functionality
+4. **Testability**: Pure CRUD functions easier to test in isolation
+5. **Scalability**: Clear patterns for future module development
+
+## Files Modified
+
+### New Files Created:
+- `src/modules/measure/application/usecases/measureState.ts`
+- `src/modules/diet/food/application/usecases/foodCrud.ts`
+
+### Files Modified:
+- `src/modules/measure/application/usecases/measureCrud.ts`
+- `src/modules/diet/food/application/food.ts`
+- `src/modules/measure/infrastructure/supabase/realtime.ts`
+- `src/sections/profile/measure/components/BodyMeasureView.tsx`
+- `src/sections/profile/measure/components/BodyMeasuresEvolution.tsx`
+
+## Migration Complete
+
+All recently refactored modules now consistently follow the usecase pattern established in day-diets. The codebase maintains architectural consistency while preserving all existing functionality.
