+++ src/sections/macro-nutrients/components/MacroTargets.tsx
@@ -11,7 +11,7 @@ import {
 import {
   insertMacroProfile,
   updateMacroProfile,
-} from '~/modules/diet/macro-profile/application/usecases/macroProfileState'
+} from '~/modules/diet/macro-profile/application/usecases/macroProfileCrud'
 import {
   createNewMacroProfile,
   type MacroProfile,
